+++ axes/apps.py
@@ -1,7 +1,10 @@
+# pylint: disable=import-outside-toplevel, unused-import
+
 from logging import getLogger
 
 from django import apps
-from pkg_resources import get_distribution
+
+from axes import __version__
 
 log = getLogger(__name__)
 
@@ -25,22 +28,29 @@ class AppConfig(apps.AppConfig):
         cls.initialized = True
 
         # Only import settings, checks, and signals one time after Django has been initialized
-        from axes.conf import settings  # noqa
-        from axes import checks, signals  # noqa
+        from axes.conf import settings
+        from axes import checks, signals
 
         # Skip startup log messages if Axes is not set to verbose
         if settings.AXES_VERBOSE:
-            if settings.AXES_ONLY_USER_FAILURES:
-                mode = "blocking by username only"
-            elif settings.AXES_LOCK_OUT_BY_COMBINATION_USER_AND_IP:
-                mode = "blocking by combination of username and IP"
-            elif settings.AXES_LOCK_OUT_BY_USER_OR_IP:
-                mode = "blocking by username or IP"
+            if callable(settings.AXES_LOCKOUT_PARAMETERS) or isinstance(
+                settings.AXES_LOCKOUT_PARAMETERS, str
+            ):
+                mode = "blocking by parameters that are calculated in a custom callable"
+
             else:
-                mode = "blocking by IP only"
+                mode = "blocking by " + " or ".join(
+                    [
+                        param
+                        if isinstance(param, str)
+                        else "combination of " + " and ".join(param)
+                        for param in settings.AXES_LOCKOUT_PARAMETERS
+                    ]
+                )
+
             log.info(
                 "AXES: BEGIN version %s, %s",
-                get_distribution("django-axes").version,
+                __version__,
                 mode,
             )