+++ axes/migrations/0002_auto_20151217_2044.py
@@ -2,7 +2,6 @@ from django.db import migrations, models
 
 
 class Migration(migrations.Migration):
-
     dependencies = [("axes", "0001_initial")]
 
     operations = [